# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Media Hub - Multi-store clothing management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface ChangePasswordInput {
  username?: string;
  currentPassword?: string;
  newPassword?: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  super_admin: 'super_admin',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  username: string;
  role: UserRole;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface AdminInput {
  username: string;
  password: string;
}

export type ProductProductType = typeof ProductProductType[keyof typeof ProductProductType];


export const ProductProductType = {
  Top: 'Top',
  Bottom: 'Bottom',
  Full_Outfit: 'Full Outfit',
  Functional: 'Functional',
} as const;

export interface Product {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  images: string[];
  discountPrice: number;
  actualPrice: number;
  /** @nullable */
  functionCategory?: string | null;
  productType: ProductProductType;
  sizes: string[];
  likeCount: number;
  /** @nullable */
  storeId?: string | null;
  createdAt: string;
}

export type ProductInputProductType = typeof ProductInputProductType[keyof typeof ProductInputProductType];


export const ProductInputProductType = {
  Top: 'Top',
  Bottom: 'Bottom',
  Full_Outfit: 'Full Outfit',
  Functional: 'Functional',
} as const;

export interface ProductInput {
  name: string;
  description?: string;
  images: string[];
  discountPrice: number;
  actualPrice: number;
  functionCategory?: string;
  productType: ProductInputProductType;
  sizes: string[];
}

export type ProductUpdateProductType = typeof ProductUpdateProductType[keyof typeof ProductUpdateProductType];


export const ProductUpdateProductType = {
  Top: 'Top',
  Bottom: 'Bottom',
  Full_Outfit: 'Full Outfit',
  Functional: 'Functional',
} as const;

export interface ProductUpdate {
  name?: string;
  description?: string;
  images?: string[];
  discountPrice?: number;
  actualPrice?: number;
  functionCategory?: string;
  productType?: ProductUpdateProductType;
  sizes?: string[];
}

export interface LikeResponse {
  likeCount: number;
}

export interface ImageUploadInput {
  /** Base64 encoded image data */
  imageData: string;
  fileName: string;
}

export interface ImageUploadResponse {
  url: string;
}

export interface Category {
  id: string;
  name: string;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
}

export interface Booking {
  id: string;
  productId: string;
  product: Product;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  selectedSize: string;
  ignored: boolean;
  createdAt: string;
}

export interface BookingInput {
  productId: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  selectedSize: string;
}

export interface Store {
  id: string;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  openingTime?: string | null;
  /** @nullable */
  openDays?: string | null;
  /** @nullable */
  bannerImage?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  publicSlug?: string | null;
  isLocked: boolean;
  createdAt?: string;
}

export interface StoreInput {
  name: string;
  address?: string;
  whatsappNumber?: string;
  openingTime?: string;
  openDays?: string;
  bannerImage?: string;
  description?: string;
}

export interface StoreUpdate {
  name?: string;
  address?: string;
  whatsappNumber?: string;
  openingTime?: string;
  openDays?: string;
  bannerImage?: string;
  description?: string;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  booking: 'booking',
  like: 'like',
  store_request: 'store_request',
} as const;

export interface Notification {
  id: string;
  type: NotificationType;
  message: string;
  read: boolean;
  /** @nullable */
  relatedId?: string | null;
  createdAt: string;
}

export type DashboardSummaryCategoryCounts = {
  Top: number;
  Bottom: number;
  'Full Outfit': number;
  Functional: number;
};

export interface DashboardSummary {
  totalProducts: number;
  categoryCounts: DashboardSummaryCategoryCounts;
  activeBookings: number;
  unreadNotifications: number;
  recentBookings?: Booking[];
}

export type ListProductsParams = {
type?: ListProductsType;
storeId?: string;
};

export type ListProductsType = typeof ListProductsType[keyof typeof ListProductsType];


export const ListProductsType = {
  Top: 'Top',
  Bottom: 'Bottom',
  Full_Outfit: 'Full Outfit',
  Functional: 'Functional',
} as const;