# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Web Media Hub - Multi-store clothing management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/change-password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change username and/or password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all function/event categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new function/event category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all active bookings (newest first)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (public, for customers)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/ignore:
    patch:
      operationId: ignoreBooking
      tags: [bookings]
      summary: Ignore (soft-delete) a booking
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking ignored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List all admins (super_admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admins list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createAdmin
      tags: [admins]
      summary: Create a new admin (super_admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admins/{id}:
    delete:
      operationId: deleteAdmin
      tags: [admins]
      summary: Delete an admin (super_admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    ChangePasswordInput:
      type: object
      properties:
        username:
          type: string
        currentPassword:
          type: string
        newPassword:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [super_admin, admin]
        createdAt:
          type: string
    AdminInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    Product:
      type: object
      required: [id, name, discountPrice, actualPrice, productType, sizes, images, likeCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        discountPrice:
          type: number
        actualPrice:
          type: number
        functionCategory:
          type: ["string", "null"]
        productType:
          type: string
          enum: [Top, Bottom, "Full Outfit", Functional]
        sizes:
          type: array
          items:
            type: string
        likeCount:
          type: integer
        storeId:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Category:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    Booking:
      type: object
      required: [id, productId, product, customerName, customerPhone, customerAddress, selectedSize, createdAt, ignored]
      properties:
        id:
          type: string
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        selectedSize:
          type: string
        ignored:
          type: boolean
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [productId, customerName, customerPhone, customerAddress, selectedSize]
      properties:
        productId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        selectedSize:
          type: string
...